chore: restructure settings folder
