+++ src/pages/UserSettings/SettingsPageAccount.tsx
@@ -6,11 +6,11 @@ import { DISCORD_INVITE_URL } from 'src/constants'
 import { fields, headings } from 'src/pages/UserSettings/labels'
 import { Flex, Heading, Text } from 'theme-ui'
 
-import { ChangeEmailForm } from './ChangeEmail.form'
-import { ChangePasswordForm } from './ChangePassword.form'
-import { PatreonIntegration } from './PatreonIntegration'
+import { PatreonIntegration } from './content/fields/PatreonIntegration'
+import { ChangeEmailForm } from './content/sections/ChangeEmail.form'
+import { ChangePasswordForm } from './content/sections/ChangePassword.form'
 
-export const AccountSettingsSection = observer(() => {
+export const SettingsPageAccount = observer(() => {
   const { description, title } = fields.deleteAccount
 
   const { userStore } = useCommonStores().stores
