+++ src/pages/UserSettings/content/sections/Focus.section.test.tsx
@@ -3,7 +3,7 @@ import '@testing-library/jest-dom/vitest'
 import { ThemeProvider } from '@emotion/react'
 import { render, screen } from '@testing-library/react'
 import { getSupportedProfileTypes } from 'src/modules/profile'
-import { FocusSection } from 'src/pages/UserSettings/content/formSections/Focus.section'
+import { FocusSection } from 'src/pages/UserSettings/content/sections/Focus.section'
 import { headings } from 'src/pages/UserSettings/labels'
 import { SettingsFormProvider } from 'src/test/components/SettingsFormProvider'
 import { describe, expect, it } from 'vitest'
